Restructuring folders and files to be able to add more generic staff
Fix typo
